Split dynamic validator tests
If we reuse a built CodeChain binary in Travis, we can merge the tests
again.
