+++ test/package.json
@@ -13,7 +13,8 @@
     "start": "yarn start-short && yarn start-long",
     "start-short": "cargo build && mocha -r ts-node/register --timeout 5000 src/e2e/*.test.ts",
     "start-long": "cargo build && mocha -r ts-node/register --timeout 10000 src/e2e.long/*.test.ts",
-    "start-dyn-val": "cargo build && mocha -r ts-node/register --timeout 10000 src/e2e.dynval/*.test.ts",
+    "start-dyn-val-1": "cargo build && mocha -r ts-node/register --timeout 10000 src/e2e.dynval/1/*.test.ts",
+    "start-dyn-val-2": "cargo build && mocha -r ts-node/register --timeout 10000 src/e2e.dynval/2/*.test.ts",
     "start-release": "yarn start-short-release && yarn start-long-release",
     "start-short-release": "cargo build --release && NODE_ENV=production mocha -r ts-node/register --timeout 5000 src/e2e/*.test.ts",
     "start-long-release": "cargo build --release && NODE_ENV=production mocha -r ts-node/register --timeout 10000 src/e2e.long/*.test.ts",
