+++ test/src/e2e.dynval/1/doubleVote.test.ts
@@ -20,11 +20,11 @@ import * as chaiAsPromised from "chai-as-promised";
 import * as stake from "codechain-stakeholder-sdk";
 import "mocha";
 
-import { Mock } from "../helper/mock";
-import { Step as TendermintStep } from "../helper/mock/tendermintMessage";
-import { validators } from "../../tendermint.dynval/constants";
-import { PromiseExpect } from "../helper/promise";
-import { setTermTestTimeout, withNodes } from "./setup";
+import { Mock } from "../../helper/mock";
+import { Step as TendermintStep } from "../../helper/mock/tendermintMessage";
+import { validators } from "../../../tendermint.dynval/constants";
+import { PromiseExpect } from "../../helper/promise";
+import { setTermTestTimeout, withNodes } from "../setup";
 
 chai.use(chaiAsPromised);
