+++ test/src/e2e.dynval/1/dv.n'.test.ts
@@ -21,10 +21,10 @@ import { SDK } from "codechain-sdk";
 import * as stake from "codechain-stakeholder-sdk";
 import "mocha";
 
-import { validators } from "../../tendermint.dynval/constants";
-import { faucetAddress, faucetSecret } from "../helper/constants";
-import { PromiseExpect } from "../helper/promise";
-import { findNode, setTermTestTimeout, withNodes } from "./setup";
+import { validators } from "../../../tendermint.dynval/constants";
+import { faucetAddress, faucetSecret } from "../../helper/constants";
+import { PromiseExpect } from "../../helper/promise";
+import { findNode, setTermTestTimeout, withNodes } from "../setup";
 
 chai.use(chaiAsPromised);
