+++ test/src/e2e.dynval/2/dv.nomination.test.ts
@@ -21,9 +21,14 @@ import { H512 } from "codechain-primitives/lib";
 import * as stake from "codechain-stakeholder-sdk";
 import "mocha";
 
-import { validators } from "../../tendermint.dynval/constants";
-import { PromiseExpect } from "../helper/promise";
-import { findNode, selfNominate, setTermTestTimeout, withNodes } from "./setup";
+import { validators } from "../../../tendermint.dynval/constants";
+import { PromiseExpect } from "../../helper/promise";
+import {
+    findNode,
+    selfNominate,
+    setTermTestTimeout,
+    withNodes
+} from "../setup";
 
 chai.use(chaiAsPromised);
