+++ test/src/e2e.dynval/2/jail.test.ts
@@ -20,9 +20,9 @@ import * as chaiAsPromised from "chai-as-promised";
 import * as stake from "codechain-stakeholder-sdk";
 import "mocha";
 
-import { validators } from "../../tendermint.dynval/constants";
-import { PromiseExpect } from "../helper/promise";
-import { setTermTestTimeout, withNodes } from "./setup";
+import { validators } from "../../../tendermint.dynval/constants";
+import { PromiseExpect } from "../../helper/promise";
+import { setTermTestTimeout, withNodes } from "../setup";
 
 chai.use(chaiAsPromised);
